Второй коммит

+++ stats-service/dto/src/main/java/ru/practicum/dto/ViewStatsDto.java
@@ -8,5 +8,5 @@ import lombok.*;
 public class ViewStatsDto {
     private String app;
     private String uri;
-    private Long hits;
+    private Integer hits;
 }
